clients/python/pyproject.toml
[project]
name = "stardust"
version = "0.1.0"
description = "Python client for stardust"
readme = "README.md"
authors = [
  { name = "olifog", email = "oliverfogelin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
  "pycapnp>=2.0.0",
]

[project.optional-dependencies]
embedded = []  # TODO: add embedded engine dependency
dev = [
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-setuptools; python_version < '3.13'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = [
  "src/stardust",
]
include = [
  "src/stardust/schemas/*.capnp",
]

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4
preview = true

[tool.ruff.lint]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
  "W",  # warnings
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "BLE",# flake8-blind-except
  "A",  # flake8-builtins
]
ignore = [
]

[tool.ruff.lint.isort]
known-first-party = ["stardust"]
combine-as-imports = true

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_any_generics = false
no_implicit_optional = true
check_untyped_defs = true
show_error_codes = true
pretty = true
explicit_package_bases = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = [
  "capnp",
]
ignore_missing_imports = true
